requirements.txt
sounddevice
numpy
scipy
SpeechRecognition
pyaudio
Pillow